Moved road & river handlers into a separate file
