+++ lib/IGameCallback.cpp
@@ -32,6 +32,8 @@
 #include "CPlayerState.h"
 #include "CSkillHandler.h"
 #include "ScriptHandler.h"
+#include "RoadHandler.h"
+#include "RiverHandler.h"
 #include "TerrainHandler.h"
 
 #include "serializer/Connection.h"
