+++ lib/mapping/MapFormatH3M.cpp
@@ -25,6 +25,8 @@
 #include "../mapObjects/MapObjects.h"
 #include "../VCMI_Lib.h"
 #include "../TerrainHandler.h"
+#include "../RoadHandler.h"
+#include "../RiverHandler.h"
 #include "../NetPacksBase.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
