+++ lib/registerTypes/TypesLobbyPacks.cpp
@@ -22,6 +22,8 @@
 #include "../CHeroHandler.h"
 #include "../spells/CSpellHandler.h"
 #include "../CTownHandler.h"
+#include "../RoadHandler.h"
+#include "../RiverHandler.h"
 #include "../TerrainHandler.h"
 #include "../mapping/CCampaignHandler.h"
 #include "../NetPacks.h"
