+++ mapeditor/objectbrowser.cpp
@@ -11,7 +11,6 @@
 #include "StdInc.h"
 #include "objectbrowser.h"
 #include "../lib/mapObjects/CObjectClassesHandler.h"
-#include "../lib/TerrainHandler.h"
 
 ObjectBrowserProxyModel::ObjectBrowserProxyModel(QObject *parent)
 	: QSortFilterProxyModel{parent}, terrain(ETerrainId::ANY_TERRAIN)
